arm64: Enable developers to provide their own device tree binaries
QEMU has the ability to take in bespoke *.dtb files on request with the
-dtb argument.  Developers can make use of this by setting the DTB
environment variable.

Notice that we're also NULing out the pflash BIOS flags here too.  This
is to prevent EFI from overriding/consuming (unconfirmed) any DTB file
that we provide.  The strange thing is, when the pflash flags are
present, /proc/device-tree is not.  Does this mean we're not booting
with any DT support?

Signed-off-by: Lee Jones <joneslee@google.com>